# swagger.yaml
swagger: "2.0"
info:
  description: |
    # This is an api that adds shadow to hand-drawn sketches.

    [`github`](https://github.com/qyzdao/ShadeSketch)
    [`paper`](https://arxiv.org/abs/2002.11812)
    [`API github`](https://github.com/vyvydkf628/ShadeSketch)

    ---

    How to use:

      * send your hand drawn sketch
      * get your hand drawn sketch with shadow automatically drawn
  version: "1.0.0"
  title: "Shade Sketch"

  contact:
    email: "qing3@umbc.edu"
  
  license:
    name: "Models are available under Creative Commons BY-NC 4.0 license. You can use, redistribute the models for non-commercial purposes."
host: "sketch.vyvydkf628.endpoint.ainize.ai"
schemes: 
  - https
  - http
paths:
  /sketch:
    post:
      tags:
      - "Shadow hand-drawn Sketches"
      summary: "Add your hand-drawn sketche"
      description: "[sample image1](https://raw.githubusercontent.com/qyzdao/ShadeSketch/master/val/10.png) \n [sample image2](https://raw.githubusercontent.com/qyzdao/ShadeSketch/master/val/1.png)"
      operationId: "add Image"
      consumes:
        - multipart/form-data
      parameters:
          - in: formData
            name: image
            type: file
            description: Image upload.
          - in: query
            name: direction
            required: false
            type: array
            uniqueItems: true
            description: Select the direction of light
            items:
              type: string
              enum: [110,210,310,410,510,610,710,810,120,220,320,420,520,620,720,820,130,230,330,430,530,630,730,830]
          - in: query
            name: size
            type: array
            uniqueItems: true
            description: Select image size
            items:
              type: string
              enum: [small,medium,large]
             
      produces:
        - image/jpg
        - image/png
        - image/gif
        - image/jpeg
      responses:
        200:
          description: A Image file.
          schema:
            type: file

  /sketch/makegif:
    post:
      tags:
      - "Shadow hand-drawn Sketches 20 frames"
      summary: "Add your hand-drawn sketche and make shadow with 20 frames"
      description: "[sample image1](https://raw.githubusercontent.com/qyzdao/ShadeSketch/master/val/10.png) \n [sample image2](https://raw.githubusercontent.com/qyzdao/ShadeSketch/master/val/1.png)"
      operationId: "add your image"
      consumes:
        - multipart/form-data
      parameters:
         - in: formData
           name: image
           type: file
           description: Image upload.
         - in: query
           name: size
           type: array
           uniqueItems: true
           description: Select image size
           items:
             type: string
             enum: [small,medium,large]
            
      produces:
        - image/jpg
        - image/png
        - image/gif
        - image/jpeg
      responses:
        200:
          description: A Image file.
          schema:
            type: file